first time getting to the end w new structure!

+++ ADC_DMA_1/filter/filter.h
@@ -19,7 +19,7 @@ struct Signal;
 void intfilter(float* b, uint16_t* a, uint16_t len_b, uint16_t len_a, uint32_t* x, float* y, uint16_t len_x);
 void floatfilter(float *b, uint16_t *a, uint16_t len_b, uint16_t len_a, float *x, float *y, uint16_t len_x);
 void statefloatfilter(struct Signal* input_signal,struct Signal* output_signal, float* filter);
-void array_conversion(uint32_t* int_array, float* float_array, uint16_t buffer_len);
+void array_conversion(uint16_t* int_array, float* float_array, uint16_t buffer_len);
 
 
 #endif /* FILTER_H_ */
